Somehow GCC on amd64 doesn't think that "unsigned long long" is equivalent to uint64_t, although sizes are the same. Had to fix for that.
git-svn-id: https://devel.codership.com/svn/galera/trunk@339 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galeracomm/vs/src/vsbes.cpp
@@ -15,11 +15,16 @@ static Logger& logger = Logger::instance();
 
 
 class Stats {
-    unsigned long long n_msgs_out;
-    unsigned long long bytes_out;
-    unsigned long long n_msgs_in;
-    unsigned long long bytes_in;
-    unsigned long long tlast;
+//    unsigned long long n_msgs_out;
+//    unsigned long long bytes_out;
+//    unsigned long long n_msgs_in;
+//    unsigned long long bytes_in;
+//    unsigned long long tlast;
+    uint64_t n_msgs_out;
+    uint64_t bytes_out;
+    uint64_t n_msgs_in;
+    uint64_t bytes_in;
+    uint64_t tlast;
 public:
     Stats() : n_msgs_out(0), bytes_out(0), n_msgs_in(0), bytes_in(0) {
 	struct timeval tv;
